introduce assignEmp2()
Create prototype manhours for Employee-class objects.

assignEmp2() uses a theoretical employee without rest days. Real employees with rest days are assigned under the assumption that the rest day of the real employee is just a regular day for the theoretical employee.

assignEmp() is degraded to just spending all remaining man hours of a real employee. This may be used in making a man hours data base of the un-assigned man hours of a real employee.